Write three test cases for heap allocation
The first test simply allocates values to the heap using a Box then asserts that the dereferenced values are equal to the integeres passed into Box::new(). The second test pushes every integer up to 1000 into a dynamically sized array (vector), then asserts that the iterated sum of the vector is equal to the nth partial sum to prove the values were allocated correctly to the vector. Finally, the third test creates boxes so as to exceed the heap size, which proves the allocator is reusing freed memory for subsequent allocations."