Create simple bump allocator
To begin our allocator design, we start with the simple bump allocator. A bump allocator allocates memory linearly in contiguous segments through the heap, and deallocates all at once.

The BumpAllocator struct has a next variable that keeps track of the next available block of memory. We maintain a structure of separating the new and init functions so as to maintain the structure used previously in the linked_list_allocator crate. The allocations variable counts the allocations.