Implement address alignment with bitmask
If align is a power of two, only one bit is set. We can use !(align-1) to clear all bits lower than the set bit and set the bit and all greater bits. Then the bitwise AND on any address will ensure that the new address will be a multiple of the alignment value, i.e. it will be properly aligned. Finally, to ensure that the address begins not at the last used aligned value of memory, but instead the next aligned value of memory, we add align-1 to the bitwise and operation.