Write test to demonstrate drawback of Bump Allocator
With the bump allocator, deallocated memory can only be reused once all memory is deallocated. If there is any allocated memory, none of the previously deallocated memory can be accessed, because the allocation pointer progresses ever rightward over the address space, never considering any leftward deallocated addresses for reuse.

To demonstrate this with a test, we create a single long-lived heap allocation, one that we do not deallocate. Then, we create a for loop in which we continuously allocate and deallocate on each iteration. We should want a heap allocator to reuse previously deallocated memory but since a Bump Allocator progresses rightward throughout the address space as long as some memory is allocated, it will continue to do this given that our long-lived initial heap allocation remains. Thus, we run out of memory.