Create ListNode struct for a Linked List Allocator
To improve upon the aforementioned disadvantages of bump allocation,
we explore a linked list allocator, which uses a linked list of
nodes to represent freed memory regions. This initial step involves
creating a ListNode struct with a size field and an optional pointer
to the next node. &'static mut describes an owned object behind a
pointer. The new function must be const fn for the static linked
list allocator. Our &'static mut ListNode requires a special
attribute in lib.rs since mutable references in const functions
are unstable.