Create function to add free memory region to front of linked list.
The add_free_region method implemented on the LinkedListAllocator
struct is how we push free regions onto the linked list. The method
takes an address and size and adds that memory region to the front.

We first assert that the region is large enough to store the ListNode,
then creates a new node with the size of the free region. Option::take
sets this new node's next pointer to the current head. The node is then
written to the beginning of the free region and the head is pointed
to it.