Create find region method for Linked List Allocator
Implement the find_region method for the LinkedListAllocator struct,
to search the linked list for a free region suitable to use for a
particular heap allocation.

We search the linked list iteratively, storing the current node in
a current variable and iterating with a while let construct that
loops as long as there are existing regions in the linked list.

On each iteration we check if the current region is suitable in
terms of both size and alignment with the alloc_from_region
method which has yet to be implemented.

If it is suitable, it is removed from the linked list by creating
temp variables and linking the current region's next pointer
(the current region is the region to be allocated) and assigning
it to the region after the region to be allocated. With the
linked list now rebuilt excluding the region to be allocated,
we return it in a tuple with its aligned alloc_start address.

If the current region is not suitable, we continue searching
the linked list, using a standard iterative technique of
reassigning the current variable to the current node's
next field.