alloc_from_region function to test and provide valid memory
For context, the find_region function traverses the linked list of
memory regions in search of a valid free region to use for
allocation. On each traversal, the logic for testing the validity
of a memory region is encapsulated within the alloc_from_region
function which we now define.

The alloc_from_region function uses the start and end address
to check for an overflow or if the end address is behind the end
address of the region. In either case, an error is returned to the
find_region function, which tells it to continue traversing for
other memory regions.

If neither occurs, the region is still potentially valid. The next
check performed by alloc_from_region ensures that the remaining
region is usable after allocation. To do so, the function checks
that the remainder is either the same size or larger than a
ListNode.